Autoschema

+++ geobeam/examples/shapefile_parcel_schema_detect.py
@@ -18,6 +18,7 @@ Example pipeline that loads a county parcel shape dataset into BigQuery.
 
 import apache_beam as beam
 from google.cloud import bigquery
+from google.api_core.exceptions import NotFound
 import geobeam
 from apache_beam.io.gcp.internal.clients import bigquery as beam_bigquery
 from apache_beam.io.gcp.bigquery_tools import parse_table_schema_from_json
@@ -126,19 +127,24 @@ def create_table(known_args):
     # TODO(developer): Set table_id to the ID of the table to create.
     #table_id = format(known_args.project, known_args.dataset, known_args.table)
     table_id=f"{known_args.project}.{known_args.dataset}.{known_args.table}"
-    table = client.delete_table(table_id)
+    
+    if not self.exists(table_id):
+        raise NotFoundException("Table does not exist")
 
-    bigquerySchema = []
-
-    bigqueryColumns = json.loads(schema_json)
-    for col in bigqueryColumns:
-        bigquerySchema.append(bigquery.SchemaField(col['name'], col['type']))
-
-    table = bigquery.Table(table_id, schema=bigquerySchema)
-    table = client.create_table(table)  # Make an API request.
-    print(
-        "Created table {}.{}.{}".format(table.project, table.dataset_id, table.table_id)
-    )
+    table_ref = self.client.dataset(self.dataset_id).table(table_id)
+    try:
+        self.client.delete_table(table_ref)
+    except NotFound:
+            # Ignore 404 error which may occur if table already deleted
+        table = client.delete_table(table_id)
+        bigquerySchema = []
+        bigqueryColumns = json.loads(schema_json)
+        for col in bigqueryColumns:
+            bigquerySchema.append(bigquery.SchemaField(col['name'], col['type']))
+            table = bigquery.Table(table_id, schema=bigquerySchema)
+            table = client.create_table(table)  # Make an API request.
+    except self.http_error as ex:
+        self.process_http_error(ex) 
 
 
 def run(pipeline_args, known_args):
